Add Opacity Painter mechanism to support independent colour and opacity
Take TwoScalarsToColor painter from the PointSprite renderer and use
it to blend a normal LUT with an opacity array supplied independently.
We insert this after the existing ScalarsToColors painter and before the
depth sort painter to give us the desired effect.

Add some options to enable/disable depth sorting and expose them in the
gui.