Improvements to depth sorting.
1) New depth sort filter which leaves original data untouched
- instead simply adds a depth sorted array of cell Ids and offset
into cellarray for {npts,ptIds}.
This save a huge amount of cell and cell data copying which is unnecessary
for a vertex coloured dataset.

2) Add a depth sort primitive painter which can use the depth sorted Cell
Id list generated by the sorter so that we very efficiently pass depth
order down the painter chain to the final render.

May need changes to handle per cell colours since we are using cells
which have the order changed but the cell data is still unchanged
(NB. Vertices are untouched so per vertex colours are ok).